Python Writing&Reading Files Real-Life Problem
TASK:
Your local university's Raptors fan club maintains a register of its active members on a .txt document. Every month they update the file by removing the members who are not active. You have been tasked with automating this with your Python skills.
Given the file currentMem, Remove each member with a 'no' in their Active column. Keep track of each of the removed members and append them to the exMem file. Make sure that the format of the original files in preserved

SOLVE:
This code moves inactive members from the currentMem file to the exMem file and keeps active members in the currentMem file. The code updates both files and prints their final contents.